fix: correctly stop syncing user after relay change
