+++ suite-common/suite-sync/src/storage.ts
@@ -1,31 +1,60 @@
-import { Evolu, SyncOwner, createOwnerWebSocketTransport } from '@evolu/common';
+import { Evolu, EvoluDeps, SyncOwner, createOwnerWebSocketTransport } from '@evolu/common';
+
+import { EvoluKeys } from '@suite-common/suite-types';
 
 import { AccountLabelSchema, AccountLabels } from './labeling/evolu/accountLabels';
 import { AddressLabelSchema, AddressLabels } from './labeling/evolu/addressLabels';
 import { OutputLabelSchema, OutputLabels } from './labeling/evolu/outputLabels';
 import { WalletLabelSchema, WalletLabels } from './labeling/evolu/walletLabels';
 import { Schema } from './schema';
+import { createEvoluInstance } from './storage/createEvoluInstance';
+
+type LocalFirstStorageDeps = {
+    relayUrl: string;
+    evoluKeys: EvoluKeys;
+    evoluDeps: EvoluDeps;
+};
 
 /**
  * This is intended as Wrapper around Evolu. In case we need to change Evolu for
  * something else, this is the Public API for the rest of the Suite ecosystem.
  */
 export class LocalFirstStorage {
+    /**
+     * Dispose function of the connected owner. When owner is changed
+     * (for example for RelayUrl change, this needs to be called).
+     * @private
+     */
+    #ownerDispose: () => void;
+
     #evolu: Evolu<typeof Schema>;
     accountLabels: AccountLabels;
     walletLabels: WalletLabels;
     outputLabels: OutputLabels;
     addressLabels: AddressLabels;
 
-    constructor(evolu: Evolu<typeof Schema>) {
-        this.#evolu = evolu;
+    constructor({ relayUrl, evoluDeps, evoluKeys }: LocalFirstStorageDeps) {
+        // just to satisfy TS for initialization,
+        // its then truly initialized in this.updateRelayUrl()
+        this.#ownerDispose = () => {};
+
+        this.#evolu = createEvoluInstance({
+            evoluKeys,
+            evoluDeps,
+        });
+        this.updateRelayUrl(relayUrl); // This updates the relay
+
         this.accountLabels = new AccountLabels(
-            evolu as unknown as Evolu<typeof AccountLabelSchema>,
+            this.#evolu as unknown as Evolu<typeof AccountLabelSchema>,
         );
-        this.walletLabels = new WalletLabels(evolu as unknown as Evolu<typeof WalletLabelSchema>);
-        this.outputLabels = new OutputLabels(evolu as unknown as Evolu<typeof OutputLabelSchema>);
+        this.walletLabels = new WalletLabels(
+            this.#evolu as unknown as Evolu<typeof WalletLabelSchema>,
+        );
+        this.outputLabels = new OutputLabels(
+            this.#evolu as unknown as Evolu<typeof OutputLabelSchema>,
+        );
         this.addressLabels = new AddressLabels(
-            evolu as unknown as Evolu<typeof AddressLabelSchema>,
+            this.#evolu as unknown as Evolu<typeof AddressLabelSchema>,
         );
     }
 
@@ -39,10 +68,12 @@ export class LocalFirstStorage {
             transports: [createOwnerWebSocketTransport({ url, ownerId: owner.id })],
         };
 
-        this.#evolu.useOwner(syncOwner);
+        this.#ownerDispose();
+        this.#ownerDispose = this.#evolu.useOwner(syncOwner);
     }
 
     async dispose() {
+        this.#ownerDispose();
         // Todo: reload prevets app from reloading. Evolu has this tab-reload
         //       to clear state as proper dispose is not yet implemented.
         //       However we cannot effort the tab-reload.
